chore: adjust data types in structs, remove unnecessary '_' prefixes

+++ include/arena.h
@@ -1,18 +1,18 @@
 #ifndef ARENA_H
 #define ARENA_H
 
-#include <stdint.h>
+#include <stddef.h>
 
 typedef struct {
   char *buffer;
-  uint64_t currentOffset;
-  uint64_t bufferSize;
+  size_t currentOffset;
+  size_t bufferSize;
 } Arena;
 
-void ArenaInit(Arena *arena, uint64_t buffer_size);
+void ArenaInit(Arena *arena, size_t buffer_size);
 
-void *ArenaAlloc(Arena *arena, uint64_t num_of_elems, uint64_t elem_size,
-                 uint64_t align_size);
+void *ArenaAlloc(Arena *arena, size_t num_of_elems, size_t elem_size,
+                 size_t align_size);
 
 void ArenaFree(Arena *arena);
